changed group hierarchy and modified commands/bindings
changed the group hierarchy to make tile-group a joint parent for both
dynamic-group and manual-group. Introduced manual-group, which takes
the place of tile-group. tile-group groups should not be made any
more, instead create manual-groups

added manual-group top and root maps, filled them with some specific
commands that dont make sense in a dynamic group.

changed many commands to only be valid in a manual group instead of a
tile group. As of this commit, a command restricted to tile-group will
be active in both manual and dynamic groups.

changed exchange-direction to check group type. This command makes
sense in both dynamic and manual groups, but dynamic groups need some
extra housekeeping to keep the stack in the proper state.