changed commands to check against a set of classes
commands now accept multiple classes instead of a single
class. Because there are now manual, dynamic, and floating groups, we
need a way for commands to be valid in float groups and one of the
tiling groups, but not the other. As of this commit we can specify
command validity like so: …(cmd-name manual-group float-group)…