implement customization gui
This commit adds two files and three dependencies to introduce a CLIM
based gui for customizing stumpwm settings.

The dependencies are slim and mcclim, which are used for the gui, and
defconfig, which is used for the customization backend.

the file customize.lisp houses the settings and the guts of
customization, while customize-gui.lisp holds the application frame
definition, display functions, and commands for the gui.